[backend] fix: change API endpoints based on auth previleges;[frontend] refactor: change all GET requests to hook; fix: update API endpoints
