+++ backend/src/routers/index.ts
@@ -4,15 +4,19 @@ const router=express.Router()
 import authMiddleware from "../middlewares/auth.middleware"
 import genericRouter from "./generic.router"
 import authRouter from "./auth.router"
-import userRouter from "./user.router"
-import orgRouter from "./org.router"
-import badgeRouter from "./badge.router"
+import userRouter from "./user/user.router"
+import userOrgRouter from "./user/org.router"
+import orgRouter from "./org/org.router"
+import orgBadgeRouter from "./org/badge.router"
 
 router.use(express.json())
 router.use("/",genericRouter)
 router.use("/auth",authRouter)
+
 router.use("/user", authMiddleware.userAuth,userRouter);
-router.use("/org", authMiddleware.userAuth, orgRouter);
-router.use("/org/badge", authMiddleware.userAuth,authMiddleware.orgAuth, badgeRouter);
+router.use("/user/org", authMiddleware.userAuth, userOrgRouter);
+
+router.use("/org", authMiddleware.userAuth, authMiddleware.orgAuth,orgRouter);
+router.use("/org/badge", authMiddleware.userAuth,authMiddleware.orgAuth, orgBadgeRouter);
 
 export default router
