+++ backend/src/routers/user/org.router.ts
@@ -1,10 +1,9 @@
 import express from "express";
 const router = express.Router();
-import orgController from "../controllers/org.controller";
+import orgController from "../../controllers/org.controller";
 
 router.get("/",orgController.listOrg)
 router.post("/add",orgController.createOrg)
 router.post("/login",orgController.loginOrg)
-router.get("/detail", orgController.detail);
 
 export default router
