+++ backend/src/routers/user/user.router.ts
@@ -1,6 +1,6 @@
 import express from "express";
 const router = express.Router();
-import userController from "../controllers/user.controller";
+import userController from "../../controllers/user.controller";
 
 router.get("/detail", userController.detail);
 router.get("/badge", userController.listBadge);
